Move schema checks functions to a cpp file
